# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "JobHunt OS API",
    "version": "0.1.0",
    "description": "Backend for JobHunt OS. Single-tenant.\n\n**Two invariants the UI must respect:**\n1. Every job carries `gaps` — JD requirements the candidate does *not* have. Always render them. Never let the UI imply the candidate has a gap technology.\n2. Nothing in this API sends anything to another human. Outbound actions (Phase 4) will be draft-and-approve only."
  },
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "name": "x-api-key",
        "in": "header"
      }
    },
...
  },
  "paths": {
...
    "/resumes/master": {
      "get": {
        "summary": "The master resume",
        "tags": [
          "resumes"
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "label": {
                      "type": "string"
                    },
                    "fileName": {
                      "nullable": true,
                      "type": "string"
                    },
                    "isMaster": {
                      "type": "boolean"
                    },
                    "doc": {
                      "type": "object",
                      "properties": {
                        "name": {
                          "type": "string"
                        },
                        "headline": {
                          "type": "string"
                        },
                        "contact": {
                          "type": "object",
                          "properties": {
                            "email": {
                              "type": "string"
                            },
                            "phone": {
                              "type": "string"
                            },
                            "location": {
                              "type": "string"
                            },
                            "links": {
                              "type": "array",
                              "items": {
                                "type": "string"
                              }
                            }
                          },
                          "required": [
                            "links"
                          ],
                          "additionalProperties": false
                        },
                        "summary": {
                          "type": "string"
                        },
                        "skills": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "group": {
                                "type": "string"
                              },
                              "items": {
                                "type": "array",
                                "items": {
                                  "type": "string"
                                }
                              }
                            },
                            "required": [
                              "items"
                            ],
                            "additionalProperties": false
                          }
                        },
                        "experience": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "company": {
                                "type": "string"
                              },
                              "title": {
                                "type": "string"
                              },
                              "location": {
                                "type": "string"
                              },
                              "start": {
                                "type": "string"
                              },
                              "end": {
                                "type": "string"
                              },
                              "bullets": {
                                "type": "array",
                                "items": {
                                  "type": "string"
                                }
                              }
                            },
                            "required": [
                              "company",
                              "title",
                              "bullets"
                            ],
                            "additionalProperties": false
                          }
                        },
                        "projects": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "name": {
                                "type": "string"
                              },
                              "tech": {
                                "type": "array",
                                "items": {
                                  "type": "string"
                                }
                              },
                              "bullets": {
                                "type": "array",
                                "items": {
                                  "type": "string"
                                }
                              },
                              "link": {
                                "type": "string"
                              }
                            },
                            "required": [
                              "name",
                              "tech",
                              "bullets"
                            ],
                            "additionalProperties": false
                          }
                        },
                        "education": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "institution": {
                                "type": "string"
                              },
                              "degree": {
                                "type": "string"
                              },
                              "start": {
                                "type": "string"
                              },
                              "end": {
                                "type": "string"
                              },
                              "details": {
                                "type": "string"
                              }
                            },
                            "required": [
                              "institution"
                            ],
                            "additionalProperties": false
                          }
                        },
                        "certifications": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        }
                      },
                      "required": [
                        "name",
                        "contact",
                        "skills",
                        "experience",
                        "projects",
                        "education",
                        "certifications"
                      ],
                      "additionalProperties": false
                    },
                    "readiness": {
                      "type": "object",
                      "properties": {
                        "ready": {
                          "type": "boolean"
                        },
                        "missing": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        }
                      },
                      "required": [
                        "ready",
                        "missing"
                      ],
                      "additionalProperties": false
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  },
                  "required": [
                    "id",
                    "label",
                    "fileName",
                    "isMaster",
                    "doc",
                    "readiness",
                    "createdAt",
                    "updatedAt"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {},
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
    "/applications/prepare": {
      "post": {
        "summary": "Tailor resumes and pre-answer forms for a batch of jobs",
        "tags": [
          "applications"
        ],
        "description": "For each job: reorder the master resume for the JD (never adding anything), truth-check it, render DOCX + PDF, write a cover letter for Tier 1, and resolve the form questions it can see against your facts. Nothing is submitted.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "jobIds": {
                    "maxItems": 100,
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "auto": {
                    "type": "object",
                    "properties": {
                      "tiers": {
                        "type": "array",
                        "items": {
                          "type": "integer",
                          "minimum": 1,
                          "maximum": 4
                        }
                      },
                      "limit": {
                        "type": "integer",
                        "minimum": 1,
                        "maximum": 50
                      },
                      "minScore": {
                        "type": "number"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "prepared": {
                      "type": "number"
                    },
                    "blocked": {
                      "type": "number"
                    },
                    "manual": {
                      "type": "number"
                    },
                    "needsAnswers": {
                      "type": "number"
                    },
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "jobId": {
                            "type": "string"
                          },
                          "applicationId": {
                            "type": "string"
                          },
                          "status": {
                            "type": "string"
                          },
                          "reason": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "jobId",
                          "status"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "prepared",
                    "blocked",
                    "manual",
                    "needsAnswers",
                    "items"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {},
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
    "/applications/approve": {
      "post": {
        "summary": "Batch approve: schedule submissions through the rate governor",
        "tags": [
          "applications"
        ],
        "description": "Approved applications are spaced 45–180 s apart, kept inside 08:00–23:00 IST and per-platform daily caps, then submitted by the local browser worker.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "ids": {
                    "minItems": 1,
                    "maxItems": 200,
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                },
                "required": [
                  "ids"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "approved": {
                      "type": "number"
                    },
                    "scheduled": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "platform": {
                            "type": "string"
                          },
                          "runAfter": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "platform",
                          "runAfter"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "approved",
                    "scheduled"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/applications/skip": {
      "post": {
        "summary": "Skip applications",
        "tags": [
          "applications"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "ids": {
                    "minItems": 1,
                    "maxItems": 200,
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                },
                "required": [
                  "ids"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "skipped": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "skipped"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
    "/applications/{id}/submitted": {
      "post": {
        "summary": "Record a manual application you submitted yourself",
        "tags": [
          "applications"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "ok"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {},
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
    "/applications/{id}/resume.{ext}": {
      "get": {
        "summary": "Download the tailored resume (pdf or docx)",
        "tags": [
          "applications"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "pdf",
                "docx"
              ]
            },
            "in": "path",
            "name": "ext",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://127.0.0.1:4000"
    }
  ],
  "security": [],
...
}